Get SSH workers to run within the context of Bundler
Many of the problems we've had with Hydra appear to stem from not fully
controlling the exact version of the code that is actually run at test
time. Forcing workers to execute within the context of a bundle partly
addresses that.

The other, non-code, part of this fix is the removal of potentially
conflicting versions of the Hydra gem installed on our systems. I've
gone ahead and done this on our master and all worker boxes; now the
only version of the Hydra gem that they should have access to is the one
that is vendored into vendor/gems in our app.